requirements.txt
numpy
scipy
scikit-learn
scikit-image
matplotlib
tqdm
pandas
tensorboard
ood_metrics
pkbar
sacred
sktime
xgboost
wandb
einops
Pillow
opencv-python
numba
torchinfo